Fixing memory leak for very large MP3 files
According to the MDN documentation on the Audio object (https://developer.mozilla.org/en-US/docs/Web/API/HTMLAudioElement/Audio), the memory allocated for each mp3 will be garbage collected only when the reference to the object goes away.

This commit drops the reference to the old audio object when the song changes.