Fixing the slider bar
Now that config.audio is being re-assigned every time the song changes, we need to remove event handlers from the old one and add them back on the new one on every song change.

We don't care about any event handlers except for "timeupdate" and "durationchange" because those are the ones we need in order to change the total duration and elapsed time on the slider bar. Those values are also needed for the slider code to work.