Configurable test config
To test the problem where one cannot start the player when setting the
`starting_playlist_song` to 0, I wanted to add some tests first, in
order to prove that this problem really exists. Doing so I realized,
that the current test structure is not adequate for that. Adding the
error producing config to the `tests/setup.js` file caused a lot of
tests to fail. As far as I could see due to unextected song indices when
skipping to the next or previous song.

So In this commit, I adapted the testing structure so that I can supply
my own config to it that is then only used for the tests in the current
file.

Overview of the changes:
- Added an option to supply additional configuration to the
  `initializeTestingElement` function. These additional settings are
  merged with the default ones usind `Object.assign`.
- Applied the same logic to the `resetConfig` function. For that I
  needed to convert the direct assignemnts to the `config` object, to an
  intermidiate `defaultConfig` object. Then I could merge everything
  together with `Object.assign`.