Added tests for testing playlist initialization
In this commit I added a bunch of test, that make sure the playlist
initialization works as expected. That is values set are properly taken
into account, and missing or invalid values have sensible defaults.

This is testing a bit more than just the error where using a
`starting_playlist_song` of `(int) 0` causing the player not to be
initialized properly.

I added that made sense to me. Though I might have some redundancy with
the other test. I'll address this in another commit, in case this is
considered an issue.